requirements.txt
absl-py==0.9.0
acme==1.1.0
appdirs==1.4.3
apturl==0.5.2
arandr==0.1.10
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
bleach==3.1.5
blinker==1.4
breezy==3.0.2
Brlapi==0.7.0
cachetools==4.1.0
certbot==0.40.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cloudpickle==1.4.1
colour==0.1.5
Comet==3.1.0
comet-git-pure==0.19.16
comet-ml==3.1.9
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
dulwich==0.19.15
entrypoints==0.3
everett==1.0.2
fastimport==0.9.8
filelock==3.0.12
future==0.18.2
google-auth==1.16.1
google-auth-oauthlib==0.4.1
gpg===1.13.1-unknown
grpcio==1.29.0
hidpidaemon==18.4.5
httplib2==0.14.0
hyperlink==19.0.0
hyperopt==0.2.4
i3pystatus==3.35
idna==2.8
imageio==2.8.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
josepy==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lensfun==0.3.2
louis==3.12.0
lxml==4.5.1
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mock==3.0.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.9
networkx==2.4
notebook==6.0.3
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
openshot-qt==2.4.3
packaging==20.3
pandas==1.0.4
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydbus==0.6.0
pyflatpak==0.4.0
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.16.0
PySocks==1.6.8
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-xlib==0.23
pytorch-lightning==0.7.6
pytorch-lightning-applications==0.0.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
Recoll==1.0
recollchm==0.8.4.1+git
repoman==1.2.2
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
system-service==0.3
systemd-python==234
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
terminado==0.8.3
test-tube==0.7.5
testpath==0.4.4
threadpoolctl==2.1.0
torbrowser-launcher==0.3.2
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
Twisted==20.3.0
typing-extensions==3.7.4.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xkit==0.0.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
